+++ doc/changelog.rst
@@ -12,6 +12,8 @@ Unreleased
 Core
 ~~~~
 
+- Add ``OrderSourceModifierModule`` interface for modifying order source
+- Add ``DiscountModule`` interface for pricing
 - Simplify ``PricingContext`` and require shop and customer for it
 - Add ``get_price_info`` etc. functions to ``shoop.core.pricing``
 - Add "codes" API to OrderSource and BaseBasket